Fixed Tiling Issue
Drawing is now done on a transparent plane (Alpha) instead of the gravel
pit. The gravel pit can now be tiled without affecting drawing with the
rake.
