+++ Assets/GravelPit/Alpha.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: e969ce170d7f13c45bc1d23261dedce3
+timeCreated: 1473234643
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
