+++ Assets/GravelPit/AlphaMat.mat.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: fca9450eef46fce4c8c4954d0f38af28
-timeCreated: 1472805171
+guid: 8ea65fdf302644c46b2704d7620adc93
+timeCreated: 1473322878
 licenseType: Free
 NativeFormatImporter:
   userData:
