+++ Assets/GravelPit/TL_Gravel_01/Material/TL_Gravel_01.mat
@@ -25,7 +25,7 @@ Material:
         name: _MainTex
       second:
         m_Texture: {fileID: 2800000, guid: 10cc3a5baaafc3f45b6c8fbe79cf4a9a, type: 3}
-        m_Scale: {x: 1, y: 1}
+        m_Scale: {x: 5, y: 5}
         m_Offset: {x: 0, y: 0}
     m_Floats: []
     m_Colors:
