+++ Assets/Resources/GravelPit/Gravel.prefab.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 5e5cc9955d5c8864cb68713c4b5a75da
-timeCreated: 1473211787
+guid: 6ddb58e5bc8f233478b015281ca69e15
+timeCreated: 1473325796
 licenseType: Free
 NativeFormatImporter:
   userData:
