--- Assets/Shaders.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: 494cc6c23d53c394a9902d5a00606a02
-folderAsset: yes
-timeCreated: 1472792396
-licenseType: Free
-DefaultImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
